+++ packages/app/package.json
@@ -41,7 +41,7 @@
     "geojson": "^0.5.0",
     "geometric": "^2.2.6",
     "lodash": "^4.17.20",
-    "lodash-es": "4.17.15",
+    "lodash-es": "^4.17.21",
     "match-sorter": "^6.2.0",
     "next": "^10.2.3",
     "polished": "^4.1.3",
@@ -111,6 +111,7 @@
     "lint-staged": "^11.0.0",
     "lodash-webpack-plugin": "^0.11.5",
     "next-compose-plugins": "^2.2.0",
+    "next-transpile-modules": "^7.3.0",
     "postcss-flexbugs-fixes": "^5.0.2",
     "postcss-preset-env": "^6.7.0",
     "prettier": "^2.3.1",
@@ -127,11 +128,12 @@
   "scripts": {
     "analyze": "cross-env ANALYZE=true yarn build",
     "export": "next export",
-    "dev": "node server/server.js",
+    "dev": "node server.mjs",
     "build": "cross-env NEXT_TELEMETRY_DISABLED=1 NODE_OPTIONS='-r esm' && next build",
     "build:nl": "cross-env NEXT_PUBLIC_LOCALE=nl && yarn build",
     "build:en": "cross-env NEXT_PUBLIC_LOCALE=en && yarn build",
-    "start": "next start -H 0.0.0.0 -p 8080",
+    "//start": "next start -H 0.0.0.0 -p 8080",
+    "start": "cross-env NODE_ENV=production node -r esm server/server.js",
     "start-e2e": "next start -H 0.0.0.0 -p 3000",
     "test": "jest --watch",
     "test:ci": "jest --ci",
