+++ packages/cms/package.json
@@ -34,7 +34,7 @@
     "download": "^8.0.0",
     "flat": "^5.0.2",
     "fs-extra": "^10.0.0",
-    "lodash-es": "4.17.15",
+    "lodash-es": "^4.17.21",
     "p-queue": "^7.1.0",
     "prompts": "^2.4.0",
     "prop-types": "^15.6",
