+++ packages/common/package.json
@@ -20,7 +20,7 @@
   },
   "dependencies": {
     "geojson": "^0.5.0",
-    "lodash-es": "4.17.15",
+    "lodash-es": "^4.17.21",
     "ts-is-present": "^1.2.1"
   },
   "devDependencies": {
