+++ tsconfig.base.json
@@ -1,6 +1,6 @@
 {
   "compilerOptions": {
-    "module": "es2020",
+    "module": "esnext",
     "lib": ["esnext", "dom"],
     "target": "es2018",
     "moduleResolution": "node",
